Byte transfer scenario crafted that should pump data through the system and see how fast it comes out the other end

+++ src/oh/DistributionPingScenario.hpp
@@ -5,7 +5,7 @@
 namespace CBR {
 class ScenarioFactory;
 class DistributionPingScenario : public Scenario {
-    size_t mNumPingsPerSecond;
+    double mNumPingsPerSecond;
     ObjectHostContext*mContext;
     Poller* mPingPoller;
     unsigned int mPingID;
